Restructuracio de fitxers i carpetes Optimitzacio reader i writer. Schema pel reader

+++ EdaCAMLibrary/Gerber/Builder/Apertures/CircleAperture.cs
@@ -44,11 +44,9 @@
             sb.Append("%ADD");
             sb.AppendFormat("{0}", Id);
             sb.Append("C,");
-            sb.AppendFormat(CultureInfo.InvariantCulture, "{0}", 
-                (double)diameter / 1000000.0);
+            sb.AppendFormat(CultureInfo.InvariantCulture, "{0}", diameter / 1000000.0);
             if (drill > 0)
-                sb.AppendFormat(CultureInfo.InvariantCulture, "X{0}", 
-                    (double)drill / 1000000.0);
+                sb.AppendFormat(CultureInfo.InvariantCulture, "X{0}", drill / 1000000.0);
             sb.Append("*%");
 
             return sb.ToString();
